requirements.txt
-i https://pypi.org/simple
alembic==1.13.2; python_version >= '3.8'
aniso8601==9.0.1
app==0.0.1
async-timeout==4.0.3; python_full_version < '3.11.3'
bcrypt==4.2.0; python_version >= '3.7'
bidict==0.23.1; python_version >= '3.8'
blinker==1.8.2; python_version >= '3.8'
certifi==2024.7.4; python_version >= '3.6'
charset-normalizer==3.3.2; python_full_version >= '3.7.0'
click==8.1.7; python_version >= '3.7'
exceptiongroup==1.2.2; python_version < '3.11'
execnet==2.1.1; python_version >= '3.8'
faker==26.3.0; python_version >= '3.8'
flask==3.0.3; python_version >= '3.8'
flask-bcrypt==1.0.1
flask-cors==4.0.1
flask-jwt-extended==4.6.0; python_version >= '3.7' and python_version < '4'
flask-mail==0.10.0; python_version >= '3.8'
flask-mailman==1.1.1; python_version >= '3.7' and python_version < '4.0'
flask-migrate==4.0.7; python_version >= '3.6'
flask-restful==0.3.10
flask-socketio==5.3.6; python_version >= '3.6'
flask-sqlalchemy==3.1.1; python_version >= '3.8'
flask-sse==1.0.0
flask-swagger-ui==4.11.1
flask-testing==0.8.1
greenlet==3.0.3; python_version < '3.13' and platform_machine == 'aarch64' or (platform_machine == 'ppc64le' or (platform_machine == 'x86_64' or (platform_machine == 'amd64' or (platform_machine == 'AMD64' or (platform_machine == 'win32' or platform_machine == 'WIN32')))))
gunicorn==23.0.0; python_version >= '3.7'
h11==0.14.0; python_version >= '3.7'
idna==3.7; python_version >= '3.5'
importlib-metadata==8.2.0; python_version < '3.10'
importlib-resources==6.4.0; python_version >= '3.8'
iniconfig==2.0.0; python_version >= '3.7'
itsdangerous==2.2.0; python_version >= '3.8'
jinja2==3.1.4; python_version >= '3.7'
mako==1.3.5; python_version >= '3.8'
markupsafe==2.1.5; python_version >= '3.7'
packaging==24.1; python_version >= '3.8'
pluggy==1.5.0; python_version >= '3.8'
psycopg2-binary==2.9.9; python_version >= '3.7'
pyjwt==2.9.0; python_version >= '3.8'
pytest==8.3.2; python_version >= '3.8'
pytest-xdist==3.6.1; python_version >= '3.8'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==1.0.1; python_version >= '3.8'
python-engineio==4.9.1; python_version >= '3.6'
python-socketio==5.11.3; python_version >= '3.8'
pytz==2024.1
redis==5.0.8; python_version >= '3.7'
requests==2.32.3; python_version >= '3.8'
simple-websocket==1.0.0; python_version >= '3.6'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlalchemy==2.0.32; python_version >= '3.7'
sqlalchemy-serializer==1.4.12; python_version >= '3.10' and python_version < '4.0'
tomli==2.0.1; python_version < '3.11'
typing-extensions==4.12.2; python_version >= '3.8'
urllib3==2.2.2; python_version >= '3.8'
werkzeug==3.0.3; python_version >= '3.8'
wsproto==1.2.0; python_full_version >= '3.7.0'
zipp==3.20.0; python_version < '3.10'